feat: separate model definition
